+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 1.0.0 (01/01/04)
+
+- updates build scripts
+- cleans up logic
+- ensures strings encoded with saywhat have no problem being decoded with the introduction of zero-width unicode spacers
+
 # 0.4.1
 
 - cleans up main file, adds named functions